sb_matching_server/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-mysql==3.9
django-cors-headers==3.5.0
psycopg2